requirements.txt
pandas
numpy
scikit-learn
yfinance
torch
pyyaml
tqdm
matplotlib
streamlit
